move integration labels inside the form and fix validation
