+++ grafana-plugin/src/containers/IntegrationForm/IntegrationForm.config.ts
@@ -1,25 +1,27 @@
 import { FormItem, FormItemType } from 'components/GForm/GForm.types';
 import { generateAssignToTeamInputDescription } from 'utils/consts';
 
+import { IntegrationFormFieldName } from './IntegrationForm.types';
+
 export const form: { name: string; fields: FormItem[] } = {
   name: 'Integration',
   fields: [
     {
       label: 'Name',
-      name: 'verbal_name',
+      name: IntegrationFormFieldName.VerbalName,
       type: FormItemType.Input,
       placeholder: 'Integration Name',
       validation: { required: true },
     },
     {
       label: 'Description',
-      name: 'description_short',
+      name: IntegrationFormFieldName.DescriptionShort,
       type: FormItemType.TextArea,
       placeholder: 'Integration Description',
     },
     {
-      name: 'team',
       label: 'Assign to team',
+      name: IntegrationFormFieldName.Team,
       description: generateAssignToTeamInputDescription('Integrations'),
       type: FormItemType.GSelect,
       extra: {
@@ -31,19 +33,24 @@ export const form: { name: string; fields: FormItem[] } = {
       },
     },
     {
-      name: 'alert_manager',
+      name: IntegrationFormFieldName.Alerting,
+      type: FormItemType.Other,
+      render: true,
+    },
+    {
+      name: IntegrationFormFieldName.AlertManager,
       type: FormItemType.Other,
     },
     {
-      name: 'contact_point',
+      name: IntegrationFormFieldName.ContactPoint,
       type: FormItemType.Other,
     },
     {
-      name: 'is_existing',
+      name: IntegrationFormFieldName.IsExisting,
       type: FormItemType.Other,
     },
     {
-      name: 'alerting',
+      name: IntegrationFormFieldName.Labels,
       type: FormItemType.Other,
       render: true,
     },
